Abstracted the logic for GUI to a separate class.
Currently, Connor.java contains both the GUI and Connor methods.

Not only is this messy, both serve different functions.

Abstracted javaFX methods and methods to separate class in gui package.
Merged the functions of Launcher into Connor.

Launcher no longer serves a purpose can be merged with Connor.
Connor now "launches" the gui.

Let's abstract away details from the main method in Gui.java.
