--- src/main/java/connor/gui/Launcher.java
@@ -1,13 +1,0 @@
-package connor.gui;
-
-import connor.Connor;
-import javafx.application.Application;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Connor.class, args);
-    }
-}
